+++ posts-app/package.json
@@ -28,6 +28,7 @@
 		"globals": "^16.2.0",
 		"typescript": "~5.8.3",
 		"typescript-eslint": "^8.34.1",
+		"typescript-plugin-css-modules": "^5.1.0",
 		"vite": "^7.0.0"
 	}
 }
