+++ posts-app/src/app/App.tsx
@@ -1,12 +1,16 @@
 import MainLayout from "../shared/layouts/MainLayout"
+import ThemeProvider from "../shared/lib/theme/ThemeProvider"
 import PostList from "../widgets/PostList/PostList"
 import "./App.css"
+import "./themes.css"
 
 function App() {
 	return (
-		<MainLayout>
-			<PostList />
-		</MainLayout>
+		<ThemeProvider>
+			<MainLayout>
+				<PostList />
+			</MainLayout>
+		</ThemeProvider>
 	)
 }
