+++ posts-app/src/shared/layouts/MainLayout.module.css
@@ -1,0 +1,5 @@
+.container {
+	margin: 0 auto;
+	max-width: 1500px;
+	width: 90%;
+}
