+++ posts-app/src/shared/lib/theme/ThemeProvider.tsx
@@ -1,34 +1,14 @@
-import React, { createContext, useContext, useEffect, useState } from "react"
+import React, { useEffect, useState } from "react"
+import { ThemeContext } from "./useTheme"
 
-import Styles from "./index.module.css"
-
-const StorageKey = "global-color-theme"
+const StorageKey = "app-theme"
 
 const supportedThemes = {
 	light: "light",
 	dark: "dark",
 }
 
-type Themes = keyof typeof supportedThemes
-
-const ThemeContext = createContext<
-	| {
-			theme: Themes
-			setTheme: (theme: Themes) => void
-			supportedThemes: { [key: string]: string }
-	  }
-	| undefined
->(undefined)
-
-const useTheme = () => {
-	const context = useContext(ThemeContext)
-
-	if (!context) {
-		throw new Error('You can use "useTheme" hook only within a <ThemeProvider> component.')
-	}
-
-	return context
-}
+export type Themes = keyof typeof supportedThemes
 
 const getTheme = (): Themes => {
 	let theme = localStorage.getItem(StorageKey)
@@ -41,12 +21,11 @@ const getTheme = (): Themes => {
 	return theme as Themes
 }
 
-const Theme = (props: { children: React.ReactNode }) => {
+const ThemeProvider = (props: { children: React.ReactNode }) => {
 	const [theme, setTheme] = useState<Themes>(getTheme)
 
 	useEffect(() => {
 		localStorage.setItem(StorageKey, theme)
-		document.documentElement.setAttribute("data-theme", theme)
 	}, [theme])
 
 	return (
@@ -62,23 +41,4 @@ const Theme = (props: { children: React.ReactNode }) => {
 	)
 }
 
-Theme.SimpleToggler = function SimpleToggler() {
-	const { theme, setTheme } = useTheme()
-
-	const handleSwitchTheme = () => {
-		if (theme === "dark") {
-			setTheme("light")
-		} else {
-			setTheme("dark")
-		}
-	}
-
-	return (
-		<div className={Styles.simpleToggler} onClick={handleSwitchTheme}>
-			<div className={Styles.ball} data-theme={theme} />
-		</div>
-	)
-}
-
-export { useTheme }
-export default Theme
+export default ThemeProvider
