+++ posts-app/tsconfig.json
@@ -7,5 +7,12 @@
 		{
 			"path": "./tsconfig.node.json"
 		}
-	]
+	],
+	"compilerOptions": {
+		"plugins": [
+			{
+				"name": "typescript-plugin-css-modules"
+			}
+		]
+	}
 }
